+++ tide-core/Cargo.toml
@@ -9,7 +9,6 @@ description = "Core types and traits for Tide web framework"
 documentation = "https://docs.rs/tide-core"
 license = "MIT OR Apache-2.0"
 repository = "https://github.com/rustasync/tide"
-autoexamples = true
 
 [dependencies]
 fnv = "1.0.6"
@@ -23,6 +22,10 @@ serde_json = "1.0.39"
 [dependencies.http-service-hyper]
 optional = true
 version = "0.2.0"
+
+[dev-dependencies]
+tide = { path = "../tide" }
+serde_derive = "1.0.91"
 
 [features]
 default = ["hyper"]
