+++ tide/Cargo.toml
@@ -19,7 +19,6 @@ repository = "https://github.com/rustasync/tide"
 version = "0.2.0"
 
 [dependencies]
-cookie = { version = "0.12", features = ["percent-encode"] }
 futures-preview = "0.3.0-alpha.16"
 http = "0.1"
 http-service = "0.2.0"
@@ -29,10 +28,8 @@ slog = "2.4.1"
 slog-async = "2.3.0"
 slog-term = "2.4.0"
 serde_urlencoded = "0.5.5"
-
-[dependencies.tide-core]
-path = "../tide-core"
-optional = true
+tide-cookies = { path = "../tide-cookies", optional = true }
+tide-core = { path = "../tide-core" }
 
 [dependencies.multipart]
 default-features = false
@@ -40,7 +37,8 @@ features = ["server"]
 version = "0.16.1"
 
 [features]
-default = ["hyper"]
+default = ["hyper", "cookies"]
+cookies = ["tide-cookies"]
 hyper = ["tide-core/http-service-hyper"]
 
 [dev-dependencies]
