+++ tide/src/middleware/mod.rs
@@ -1,6 +1,7 @@
-mod cookies;
 mod default_headers;
 mod logger;
 
-pub use self::{cookies::CookiesMiddleware, default_headers::DefaultHeaders, logger::RootLogger};
+pub use self::{default_headers::DefaultHeaders, logger::RootLogger};
+#[cfg(feature = "cookies")]
+pub use tide_cookies::CookiesMiddleware;
 pub use tide_core::middleware::{Middleware, Next};
